♻️ rename menuItem -> sidebarItem
