+++ packages/eds-lab-react/src/components/SideBar/SideBar.tokens.ts
@@ -56,7 +56,7 @@ export const sidebar: ComponentToken = {
         },
       },
     },
-    menuItem: {
+    sidebarItem: {
       typography: {
         color: expandTextColor,
       },
